dashboard/requirements.txt
dash==2.14.2
dash-ag-grid=31.0.1
dash-bootstrap-components==1.5.0
dash-daq==0.5.0
dash-extensions==1.0.4
dash-leaflet[geobuf]==0.1.28
duckdb==0.9.2
geopandas==0.14.1
pyarrow==14.0.2
python-dotenv==1.0.0
